// ui/src/app/services/models/page-response-medicine-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MedicineResponse } from '../models/medicine-response';
export interface PageResponseMedicineResponse {
  content?: Array<MedicineResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}