// ui/src/app/services/models/medicine-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MedicineRequest {
  description?: string;
  expiryDate?: string;
  manufacturer?: string;
  medicineId?: number;
  name?: string;
  price?: number;
  stockQuantity?: number;
  type?: 'SYRUP' | 'TABLETS' | 'OINTMENTS';
}